Tried implementing run_test()
I tried creating and implementing run_test() for reading test_case from .txt and test the program.
This function will read the specified test_case.txt file and it is common to all problems.
But I failed. Since The inputs and outputs of the problems are different in both data type and data structure